requirements.txt
click>=8.0.0
pandas>=1.3.0
fpdf>=1.7.2
pathlib>=1.0.1
openpyxl>=3.1.0
rich>=10.0.0
networkx>=2.6.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
alembic>=1.7.3
python-multipart>=0.0.5
pyyaml>=6.0
